fb_direct cache_adjoint: fix impossible cosine>1 from forced FP32 downcast
Yilun caught it: cosine similarity is Cauchy-Schwarz bounded by [-1,1];
1.009326 was mathematically impossible. Root cause: every gradient dict in
cache_adjoint.py force-cast to .float() (FP32) before comparison, even on
the FP64 unit test path. g_cache_direct = exact_vec - semi_vec subtracts
two highly correlated (cosine ~0.998), large vectors -- classic
catastrophic cancellation -- and doing that cancellation in FP32 instead
of native FP64 pushed the bounded quantity outside its valid range.

Removed the forced .float() casts in compute_g_exact, compute_g_semi_and_a_star,
and compute_g_cache_vjp's main dict (native dtype preserved throughout);
fixed the allow_unused zero-fallback to also match p.dtype, not just
p.device. FP64 unit test now reports cosine=1.000000, rel_norm_error=5.5e-15
(true machine precision) instead of the previous impossible value.

The real-checkpoint result (median cosine 0.940) was already computed
natively in FP32 throughout and is unaffected by this bug -- this only
corrects the toy-model sanity-check's own reporting.

+++ projects/masked-EqM/fb_direct/cache_adjoint.py
@@ -151,7 +151,7 @@ def compute_g_exact(theta, xt, t, y, ut):
         theta.zero_grad(set_to_none=True)
         loss.backward()
         g_exact = {
-            name: p.grad.detach().reshape(-1).float().clone()
+            name: p.grad.detach().reshape(-1).clone()
             for name, p in theta.named_parameters()
             if p.grad is not None
         }
@@ -222,7 +222,7 @@ def compute_g_semi_and_a_star(fb_trainer, xt, t, y, ut):
                 continue
             p = phi_named[e.backward_name]
             if p.grad is not None:
-                g_semi[e.forward_name] = p.grad.detach().reshape(-1).float().clone()
+                g_semi[e.forward_name] = p.grad.detach().reshape(-1).clone()
         loss_fb_val = float(loss_fb.detach())
 
     for p in phi.parameters():
@@ -275,8 +275,8 @@ def compute_g_cache_vjp(theta, xt, t, y, a_star, compute_per_tensor_contribution
             retain_graph=False, allow_unused=True,
         )
         g_cache_vjp = {
-            name: (g.detach().reshape(-1).float().clone() if g is not None
-                   else torch.zeros(p.numel(), device=p.device))
+            name: (g.detach().reshape(-1).clone() if g is not None
+                   else torch.zeros(p.numel(), device=p.device, dtype=p.dtype))
             for name, p, g in zip(theta_names, theta_params, grads)
         }
